Extracted getCases
